+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed issue with `fixIdStructureMismatchDocuments` maintenance function
 * `delimiter` is now the default ID structure
 * Added maintenance function to add base document ref to translated documents
+* Removed the option for overriding UI messages
 
 <!-- // @TODO remove old changelog? -->
 ## v5.6.0